// src/data/protobuf/summaryfile.ts
export interface SummaryValues {
  values: string[];
}

/** SummaryRows contains a map of variant (VariantKey) to values */
export interface SummaryRows {
  header: string[];
  /** key is VariantKey: "chrom\tpos\tref\talt" */
  rows: Map<string, SummaryValues>;
}

export interface SummaryRows_RowsEntry {
  key: string;
  value: SummaryValues | undefined;
}

/** SummaryFile represents the complete summary file structure */
export interface SummaryFile {
  rows: SummaryRows[];
}
// ...
export const SummaryValues: MessageFns<SummaryValues> = {
  encode(message: SummaryValues, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.values) {
      writer.uint32(10).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SummaryValues {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSummaryValues();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.values.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SummaryValues {
    return {
      values: globalThis.Array.isArray(object?.values) ? object.values.map((e: any) => globalThis.String(e)) : [],
    };
  },

  toJSON(message: SummaryValues): unknown {
    const obj: any = {};
    if (message.values?.length) {
      obj.values = message.values;
    }
    return obj;
  },
};

function createBaseSummaryRows(): SummaryRows {
  return { header: [], rows: new Map() };
}

export const SummaryRows: MessageFns<SummaryRows> = {
  encode(message: SummaryRows, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.header) {
      writer.uint32(10).string(v!);
    }
    message.rows.forEach((value, key) => {
      SummaryRows_RowsEntry.encode({ key: key as any, value }, writer.uint32(18).fork()).join();
    });
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SummaryRows {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSummaryRows();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.header.push(reader.string());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          const entry2 = SummaryRows_RowsEntry.decode(reader, reader.uint32());
          if (entry2.value !== undefined) {
            message.rows.set(entry2.key, entry2.value);
          }
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SummaryRows {
    return {
      header: globalThis.Array.isArray(object?.header) ? object.header.map((e: any) => globalThis.String(e)) : [],
      rows: isObject(object.rows)
        ? Object.entries(object.rows).reduce<Map<string, SummaryValues>>((acc, [key, value]) => {
          acc.set(key, SummaryValues.fromJSON(value));
          return acc;
        }, new Map())
        : new Map(),
    };
  },

  toJSON(message: SummaryRows): unknown {
    const obj: any = {};
    if (message.header?.length) {
      obj.header = message.header;
    }
    if (message.rows?.size) {
      obj.rows = {};
      message.rows.forEach((v, k) => {
        obj.rows[k] = SummaryValues.toJSON(v);
      });
    }
    return obj;
  },
};

function createBaseSummaryRows_RowsEntry(): SummaryRows_RowsEntry {
  return { key: "", value: undefined };
}

export const SummaryRows_RowsEntry: MessageFns<SummaryRows_RowsEntry> = {
  encode(message: SummaryRows_RowsEntry, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.key !== "") {
      writer.uint32(10).string(message.key);
    }
    if (message.value !== undefined) {
      SummaryValues.encode(message.value, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SummaryRows_RowsEntry {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSummaryRows_RowsEntry();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.key = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.value = SummaryValues.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SummaryRows_RowsEntry {
    return {
      key: isSet(object.key) ? globalThis.String(object.key) : "",
      value: isSet(object.value) ? SummaryValues.fromJSON(object.value) : undefined,
    };
  },

  toJSON(message: SummaryRows_RowsEntry): unknown {
    const obj: any = {};
    if (message.key !== "") {
      obj.key = message.key;
    }
    if (message.value !== undefined) {
      obj.value = SummaryValues.toJSON(message.value);
    }
    return obj;
  },
};

function createBaseSummaryFile(): SummaryFile {
  return { rows: [] };
}

export const SummaryFile: MessageFns<SummaryFile> = {
  encode(message: SummaryFile, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.rows) {
      SummaryRows.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SummaryFile {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSummaryFile();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.rows.push(SummaryRows.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SummaryFile {
    return { rows: globalThis.Array.isArray(object?.rows) ? object.rows.map((e: any) => SummaryRows.fromJSON(e)) : [] };
  },

  toJSON(message: SummaryFile): unknown {
    const obj: any = {};
    if (message.rows?.length) {
      obj.rows = message.rows.map((e) => SummaryRows.toJSON(e));
    }
    return obj;
  },
};

function isObject(value: any): boolean {
  return typeof value === "object" && value !== null;
}

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
}